C names of files to uppercase
