+++ src/BishopClass.js
@@ -13,48 +13,30 @@ var __extends = (this && this.__extends) || (function () {
     };
 })();
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.Queen = void 0;
-var Piece_1 = require("./Piece");
+exports.Bishop = void 0;
+var PieceClass_1 = require("./PieceClass");
 var logic_1 = require("./logic");
 var logic_2 = require("./logic");
-var Queen = /** @class */ (function (_super) {
-    __extends(Queen, _super);
-    function Queen(color, location) {
-        var _this = this;
-        if (color == 'white')
-            location = [4, 1];
-        else
-            location = [4, 8];
-        _this = _super.call(this, 'queen', color, location) || this;
-        return _this;
+var Bishop = /** @class */ (function (_super) {
+    __extends(Bishop, _super);
+    function Bishop(color, location) {
+        return _super.call(this, 'bishop', color, location) || this;
     }
-    Queen.prototype.checkPossibleMoves = function () {
+    Bishop.prototype.checkPossibleMoves = function () {
         this.possibleLocations = [];
-        for (var i = 0; i < 8; i++) {
+        for (var i = 0; i < 4; i++) {
             var dir = [0, 0];
             switch (i) {
                 case 0:
-                    dir = [1, 0];
-                    break;
-                case 1:
-                    dir = [-1, 0];
-                    break;
-                case 2:
-                    dir = [0, 1];
-                    break;
-                case 3:
-                    dir = [0, -1];
-                    break;
-                case 4:
                     dir = [1, 1];
                     break;
-                case 5:
+                case 1:
                     dir = [-1, 1];
                     break;
-                case 6:
+                case 2:
                     dir = [1, -1];
                     break;
-                case 7:
+                case 3:
                     dir = [-1, -1];
                     break;
             }
@@ -66,9 +48,9 @@ var Queen = /** @class */ (function (_super) {
                     checkingPosition[1] > 8)
                     break;
                 var PieceOnWay = logic_2.AREASARRAY[logic_1.getAreaArrayIndex(checkingPosition)].piece;
-                if (PieceOnWay instanceof Piece_1.Piece && PieceOnWay.color == this.color)
+                if (PieceOnWay instanceof PieceClass_1.Piece && PieceOnWay.color == this.color)
                     break;
-                else if (PieceOnWay instanceof Piece_1.Piece && PieceOnWay.color !== this.color) {
+                else if (PieceOnWay instanceof PieceClass_1.Piece && PieceOnWay.color !== this.color) {
                     this.possibleLocations.push(checkingPosition);
                     break;
                 }
@@ -76,6 +58,6 @@ var Queen = /** @class */ (function (_super) {
             }
         }
     };
-    return Queen;
-}(Piece_1.Piece));
-exports.Queen = Queen;
+    return Bishop;
+}(PieceClass_1.Piece));
+exports.Bishop = Bishop;
