+++ src/KnightClass.ts
@@ -1,4 +1,4 @@
-import { Piece, coordinates } from './Piece';
+import { Piece, coordinates } from './PieceClass';
 import { getAreaArrayIndex } from './logic';
 import { AREASARRAY } from './logic';
