+++ src/logic.js
@@ -1,13 +1,13 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.getAreaArrayIndex = exports.listenDOMchessboard = exports.changeArrayCoordinatesToString = exports.createBoardArray = exports.AREASARRAY = void 0;
-var Piece_1 = require("./Piece");
-var King_1 = require("./King");
-var Queen_1 = require("./Queen");
-var Rook_1 = require("./Rook");
-var Bishop_1 = require("./Bishop");
-var Pawn_1 = require("./Pawn");
-var Knight_1 = require("./Knight");
+var PieceClass_1 = require("./PieceClass");
+var KingClass_1 = require("./KingClass");
+var QueenClass_1 = require("./QueenClass");
+var RookClass_1 = require("./RookClass");
+var BishopClass_1 = require("./BishopClass");
+var PawnClass_1 = require("./PawnClass");
+var KnightClass_1 = require("./KnightClass");
 exports.AREASARRAY = [];
 var BOARD = document.querySelector('.board');
 var Letters;
@@ -49,25 +49,25 @@ var createBoardArray = function () {
 };
 exports.createBoardArray = createBoardArray;
 var insertPieces = function () {
-    var kingW = new King_1.King('white');
-    var kingB = new King_1.King('black');
-    var queenW = new Queen_1.Queen('white');
-    var queenB = new Queen_1.Queen('black');
-    var rockW1 = new Rook_1.Rook('white', [1, 1]);
-    var rockW2 = new Rook_1.Rook('white', [8, 1]);
-    var rockB1 = new Rook_1.Rook('black', [1, 8]);
-    var rockB2 = new Rook_1.Rook('black', [8, 8]);
-    var bishopW1 = new Bishop_1.Bishop('white', [3, 1]);
-    var bishopW2 = new Bishop_1.Bishop('white', [6, 1]);
-    var bishopB1 = new Bishop_1.Bishop('black', [3, 8]);
-    var bishopB2 = new Bishop_1.Bishop('black', [6, 8]);
-    var knightW1 = new Knight_1.Knight('white', [2, 1]);
-    var knightW2 = new Knight_1.Knight('white', [7, 1]);
-    var knightB1 = new Knight_1.Knight('black', [2, 8]);
-    var knightB2 = new Knight_1.Knight('black', [7, 8]);
+    var kingW = new KingClass_1.King('white');
+    var kingB = new KingClass_1.King('black');
+    var queenW = new QueenClass_1.Queen('white');
+    var queenB = new QueenClass_1.Queen('black');
+    var rockW1 = new RookClass_1.Rook('white', [1, 1]);
+    var rockW2 = new RookClass_1.Rook('white', [8, 1]);
+    var rockB1 = new RookClass_1.Rook('black', [1, 8]);
+    var rockB2 = new RookClass_1.Rook('black', [8, 8]);
+    var bishopW1 = new BishopClass_1.Bishop('white', [3, 1]);
+    var bishopW2 = new BishopClass_1.Bishop('white', [6, 1]);
+    var bishopB1 = new BishopClass_1.Bishop('black', [3, 8]);
+    var bishopB2 = new BishopClass_1.Bishop('black', [6, 8]);
+    var knightW1 = new KnightClass_1.Knight('white', [2, 1]);
+    var knightW2 = new KnightClass_1.Knight('white', [7, 1]);
+    var knightB1 = new KnightClass_1.Knight('black', [2, 8]);
+    var knightB2 = new KnightClass_1.Knight('black', [7, 8]);
     for (var i = 1; i < 9; i++) {
-        var pawnW = new Pawn_1.Pawn('white', [i, 2]);
-        var pawnB = new Pawn_1.Pawn('black', [i, 7]);
+        var pawnW = new PawnClass_1.Pawn('white', [i, 2]);
+        var pawnB = new PawnClass_1.Pawn('black', [i, 7]);
     }
     console.table(exports.AREASARRAY);
 };
@@ -90,7 +90,7 @@ var listenSelection = function (e) {
     var stringCoordinates = target.classList[0];
     var arr = changeStringCoordinatesToArray(stringCoordinates);
     var index = exports.getAreaArrayIndex(arr);
-    if (exports.AREASARRAY[index].piece instanceof Piece_1.Piece) {
+    if (exports.AREASARRAY[index].piece instanceof PieceClass_1.Piece) {
         selectPiece(arr);
         BOARD.removeEventListener('click', listenSelection);
     }
@@ -111,7 +111,7 @@ var selectPiece = function (position) {
         BOARD.removeEventListener('click', listenNewPosition);
         BOARD.addEventListener('click', listenSelection);
     };
-    if (currentPiece instanceof Piece_1.Piece) {
+    if (currentPiece instanceof PieceClass_1.Piece) {
         BOARD.addEventListener('click', listenNewPosition);
     }
     else
